Simplify workspace resolver interface, rework various types
