+++ packages/modular-scripts/src/check/verifyBrowserslist.ts
@@ -5,10 +5,11 @@ import * as path from 'path';
 import prompts from 'prompts';
 
 import * as logger from '../utils/logger';
-import { ModularPackageJson } from '../utils/isModularType';
 import getModularRoot from '../utils/getModularRoot';
 import getWorkspaceInfo from '../utils/getWorkspaceInfo';
 import { defaultBrowsers } from '../utils/checkBrowsers';
+
+import type { ModularPackageJson } from 'modular-types';
 
 export async function check(): Promise<boolean> {
   let valid = true;
